fix(review): rename test target to 'sync' per feedback
Merge the integrity tests into the existing 'sync' test target at
tests/e2e/sync.rs under an  module. The e2e daemon tests
remain #[ignore]d; the integrity tests run by default. Remove the
separate sync_integrity [[test]] entry and tests/sync.rs.
